fix typo

+++ lib/BOM/Test/Script.pm
@@ -33,10 +33,10 @@ sub start_script_if_not_running {
 sub check_script {
     my $self = shift;
     my $name = $self->name;
-    return 0 unless -f $self->pid;
+    return 0 unless $self->pid;
     my $pid = $self->pid;
     system("/usr/bin/pgrep --ns $pid $name");
-    reutrn !$?;    # return  true if pgrep success
+    return !$?;    # return  true if pgrep success
 }
 
 sub start_script {
@@ -50,7 +50,7 @@ sub start_script {
 sub stop_script {
     my $self = shift;
     my $pid  = $self->pid;
-    reutrn unless $self->check_script;
+    return unless $self->check_script;
     kill 'SIGTERM' => $pid;
     wait_till_exit($pid, 10);
     return;
